+++ src/mutation/MutationWithdrawBufferToken.ts
@@ -47,7 +47,6 @@ License: https://github.com/WishKnish/KnishIO-Client-TS/blob/master/LICENSE
 */
 
 import MutationProposeMolecule from './MutationProposeMolecule'
-import type Wallet from '@/core/Wallet'
 
 /**
  * Mutation for withdrawing buffer tokens
@@ -57,15 +56,12 @@ export default class MutationWithdrawBufferToken extends MutationProposeMolecule
    * Fills the molecule with buffer withdrawal data
    */
   override fillMolecule({
-    recipients,
-    signingWallet
+    recipients
   }: {
     recipients: Record<string, any>
-    signingWallet?: Wallet
   }): void {
-    (this.$__molecule as any).initWithdrawBuffer({
-      recipients,
-      signingWallet
+    this.$__molecule.initWithdrawBuffer({
+      recipients
     })
     this.$__molecule.sign({})
     this.$__molecule.check(this.$__molecule.sourceWallet)